requirements.txt
Flask
python-dotenv
PyPDF2
python-docx
groq
pymongo
flask-login
werkzeug